rename restaurant card
